+++ Materiały pomocnicze/Pomocznicze pliki/ResourceType.kt
@@ -1,0 +1,7 @@
+package pl.krzysztof.lipka.laborki.model
+
+enum class ResourceType(val value: String) {
+
+    LAND_FORCES("LAND_FORCES"),
+    MARINE_FLEET("MARINE_FLEET"),
+}
