--- dimens.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <dimen name="icon_size_standard">24dp</dimen>
-    <dimen name="margin_small">8dp</dimen>
-    <dimen name="margin_standard">16dp</dimen>
-    <dimen name="margin_medium">24dp</dimen>
-    <dimen name="margin_double">32dp</dimen>
-    <dimen name="margin_large">64dp</dimen>
-    <dimen name="margin_very_large">128dp</dimen>
-    <dimen name="text_size_small">12sp</dimen>
-    <dimen name="coordinates_edit_text_width">105dp</dimen>
-    <dimen name="bottom_sheet_share_menu_peek_height">80dp</dimen>
-    <dimen name="address_email_edit_text_width">150dp</dimen>
-    <dimen name="email_edit_text_width">210dp</dimen>
-    <dimen name="map_height">350dp</dimen>
-</resources>
